Refactor: auth router refactoring
